Move WfssMultiSpecModel to WfssSpecModel
