+++ src/stdatamodels/jwst/datamodels/schemas/wfss_specmeta.schema.yaml
@@ -1,7 +1,7 @@
 %YAML 1.1
 ---
 $schema: "http://stsci.edu/schemas/asdf/asdf-schema-1.1.0"
-id: "http://stsci.edu/schemas/jwst_datamodel/wfss_multispecmeta.schema"
+id: "http://stsci.edu/schemas/jwst_datamodel/wfss_specmeta.schema"
 type: object
 properties:
   filename:
