+++ src/stdatamodels/jwst/datamodels/spec.py
@@ -1,7 +1,7 @@
 from .model_base import JwstDataModel
 
 
-__all__ = ["SpecModel", "MRSSpecModel", "TSOSpecModel"]
+__all__ = ["SpecModel", "MRSSpecModel", "TSOSpecModel", "WFSSSpecModel"]
 
 
 class SpecModel(JwstDataModel):
@@ -56,6 +56,32 @@ class TSOSpecModel(JwstDataModel):
     """
 
     schema_url = "http://stsci.edu/schemas/jwst_datamodel/tso_spec.schema"
+
+    def get_primary_array_name(self):  # noqa: D102
+        return "spec_table"
+
+
+class WFSSSpecModel(JwstDataModel):
+    """
+    A data model for NIRCam andNIRISS WFSS 1D spectra.
+
+    This model differs from the other MultiSpecModel classes in that
+    it is designed to hold all the spectra from a single WFSS exposure
+    in a "flat" table format. Therefore, it does not have the `spec` attribute
+    that is present in the other MultiSpecModel classes. Instead, it has
+    a `spec_table` attribute that contains the spectral data and metadata
+    for all sources in the exposure.
+
+    Attributes
+    ----------
+    spec_table : numpy table
+        Table containing the extracted spectral data for all sources in a WFSS exposure.
+        The table still has the standard spectral columns, but also has additional
+        metadata columns that are used to identify the source
+        and the spectral extraction region.
+    """
+
+    schema_url = "http://stsci.edu/schemas/jwst_datamodel/wfss_spec.schema"
 
     def get_primary_array_name(self):  # noqa: D102
         return "spec_table"
